Work out OpenId Connect setup
Currently requires the server to be restarted after setting the OpenId
Connect authority. What it expectes as roles, client names is very
hardcoded. As will stay for the time being. It should all be changeable,
but for now that's not something I'm looking to support. No extra
complexity should be added to this configuration in the future. So I can
hold off with it