Facilities
I think this should make them all createable, updateable, and deletable;
The finer logic behind each of them still has to be done. But I think
this is better done while having the UI ready; so it's easier. Instead
of plugging it in now, and then maybe having to fix other stuff while
integrating it into the UI